SistAgendamentos/requirements.txt
fastapi
uvicorn
supabase
python-dotenv
fastapi-mail
pydantic[email]
PyJWT
